+++ amatracker/Movement List View Controller/MovementListViewController.swift
@@ -2,7 +2,7 @@ import UIKit
 import CoreData
 import Mixpanel
 
-class MovementListViewController: UIViewController, UITableViewDataSource, UITableViewDelegate  {
+class MovementListViewController: UIViewController  {
     
     // MARK: Constants
     let MOVEMENT_DETAIL_SEGUE = "movementDetailSegue"
@@ -34,28 +34,6 @@ class MovementListViewController: UIViewController, UITableViewDataSource, UITab
         super.didReceiveMemoryWarning()
     }
     
-    // MARK: UITableViewDataSource functions
-    
-    func tableView(_ tableView: UITableView, cellForRowAt indexPath: IndexPath) -> UITableViewCell {
-        
-        guard let movementViewModel = viewModel.movementViewModel(for: indexPath.row) else {
-            fatalError("Could not getfind movement view model")
-        }
-        let cell: MovementTableCell = tableView.dequeueReusableCell(withIdentifier: MovementTableCell.reuseIdentifier) as! MovementTableCell
-        cell.configure(fromViewModel: movementViewModel)
-        
-        return cell
-    }
-    
-    func tableView(_ tableView: UITableView, numberOfRowsInSection section: Int) -> Int {
-        return viewModel.numberOfMovements
-    }
-    
-    // MARK: UITableViewDelegate required functions
-    func tableView(_ tableView: UITableView, didSelectRowAt indexPath: IndexPath) {
-        performSegue(withIdentifier: MOVEMENT_DETAIL_SEGUE, sender: self)
-    }
-
     // MARK: - Navigation
     
     override func prepare(for segue: UIStoryboardSegue, sender: Any?) {
@@ -63,11 +41,13 @@ class MovementListViewController: UIViewController, UITableViewDataSource, UITab
         if(segue.identifier == MOVEMENT_DETAIL_SEGUE) {
             let destinationViewController : MovementDetailViewController = segue.destination  as! MovementDetailViewController
             
-            let indexPathOfSelectedRow : IndexPath = self.movementTable.indexPathForSelectedRow!
-            let selectedMovement = viewModel.movementViewModel(for: indexPathOfSelectedRow.row)?.movement
-            destinationViewController.movement = selectedMovement
+            if let indexPathOfSelectedRow : IndexPath = self.movementTable.indexPathForSelectedRow,
+                let selectedMovement = viewModel.movementViewModel(for: indexPathOfSelectedRow.row)?.movement{
+                
+                     destinationViewController.movement = selectedMovement
+               
+            }
         }
-        
     }
     
     override var preferredStatusBarStyle : UIStatusBarStyle {
